transformer/requirements.txt
certifi==2022.12.7
chardet==4.0.0
click==8.1.3
colorama==0.4.6
cupy-cuda112==10.6.0
cycler==0.11.0
fastrlock==0.8.1
fonttools==4.38.0
h5py==3.8.0
idna==2.10
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.5.3
numba==0.56.0
numpy==1.22.4
packaging==23.0
pandas==1.5.3
perun==0.2.0
Pillow==9.4.0
prettytable==3.7.0
psutil==5.9.5
py-cpuinfo==9.0.0
py-spy==0.3.14
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.25.1
six==1.16.0
tqdm==4.64.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
